backend-radiotelescope/requirements.txt
aniso8601==9.0.1
click==8.0.3
dataclasses==0.8
Flask==2.0.2
Flask-RESTful==0.3.9
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
paho-mqtt==1.6.1
pytz==2021.3
six==1.16.0
typing_extensions==4.0.1
Werkzeug==2.0.2
zipp==3.6.0
